refactor: rename displaypanel.rs to display_panel.rs
