+++ prototypes/src/entity.rs
@@ -29,7 +29,7 @@ mod cargo_landing_pad;
 mod combinators;
 mod containers;
 mod crafting_machines;
-mod displaypanel;
+mod display_panel;
 mod electric_energy_interface;
 mod electric_pole;
 mod flying_robots;
@@ -81,7 +81,7 @@ pub use cargo_landing_pad::*;
 pub use combinators::*;
 pub use containers::*;
 pub use crafting_machines::*;
-pub use displaypanel::*;
+pub use display_panel::*;
 pub use electric_energy_interface::*;
 pub use electric_pole::*;
 pub use flying_robots::*;
