+++ prototypes/src/entity/display_panel.rs
